requirements/prod.txt
pyelasticsearch==0.4.1
requests==1.1.0
simplejson==3.1.0
statsd-client==1.0.4
